+++ src/app/app-routing-module.ts
@@ -1,8 +1,12 @@
 import { NgModule } from '@angular/core';
 import type { Routes } from '@angular/router';
 import { RouterModule } from '@angular/router';
+import { Login } from './pages/login/login';
 
-const routes: Routes = [];
+const routes: Routes = [
+  { path: 'login', component: Login },
+  { path: '', redirectTo: '/login', pathMatch: 'full' }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
