+++ src/app/app.html
@@ -8,9 +8,7 @@
 </head>
 
 <body>
-  <h1>Sistema de agendamento</h1>
-
-  <app-login></app-login>
+  <router-outlet></router-outlet>
 </body>
 
 </html>
