--- src/app/components/shared/login/login.css
@@ -1,28 +1,0 @@
-.header {
-  background-image: linear-gradient(to bottom, #870C08, #210302);
-  flex-shrink: 0;
-}
-
-.card-login {
-  background-color: #f9f9fb;
-  margin-top: -25px;
-  border-top-left-radius: 35px !important;
-  border-top-right-radius: 35px !important;
-  box-shadow: 0 -5px 15px rgba(0,0,0,0.1);
-}
-
-.btn-danger{
-    background-color: #870C08 !important;
-}
-
-.form-control{
-    border-color: #870C08 !important;
-}
-
-h1{
-    font-size: 24px !important;
-}
-
-h2{
-    font-size: 48px !important;
-}
