--- src/app/components/shared/login/login.ts
@@ -1,15 +1,0 @@
-import { Component } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
-
-@Component({
-  selector: 'app-login',
-  standalone: false,
-  templateUrl: './login.html',
-  styleUrl: './login.css'
-})
-export class Login {
-    loginForm = new FormGroup({
-      user: new FormControl(''),
-      password: new FormControl('')
-    });
-}
