+++ src/app/pages/login/login.css
@@ -1,0 +1,49 @@
+.header {
+    background-image: linear-gradient(to bottom, #870C08, #210302);
+    flex-shrink: 0;
+}
+
+.card-login {
+    background-color: #f9f9fb;
+    margin-top: -25px;
+    border-top-left-radius: 35px !important;
+    border-top-right-radius: 35px !important;
+    box-shadow: 0 -5px 15px rgba(0, 0, 0, 0.1);
+}
+
+.btn-danger {
+    background-color: #F21E18 !important;
+}
+
+.btn-danger:active {
+    background-color: #600806 !important;
+}
+
+.btn-danger:hover {
+    background-color: #AD0F0A !important;
+}
+
+.form-control {
+    border-color: #870C08 !important;
+}
+
+h1 {
+    font-size: 24px !important;
+    text-shadow: 2px 2px 4px #000000;
+}
+
+h2 {
+    font-size: 100px !important;
+    text-shadow: 4px 4px 6px #000000;
+}
+
+h3 {
+    font-size: 32px !important;
+}
+
+.b-line {
+    padding: 10px;
+    border-bottom: 1px black solid;
+    width: 50%;
+    margin: 0 auto;
+}
