+++ src/app/pages/login/login.html
@@ -1,0 +1,71 @@
+<div class="login-page d-flex flex-column min-vh-100">
+
+    <div class="header text-white py-5">
+        <img src="fatec-logo.png" alt="">
+        <div class="text-center mt-3">
+            <h1 class="fw-bold">Seja bem-vindo ao<br>Gestor de Espaços Acadêmicos</h1>
+            <h2 class="fw-bold mt-4">GEA</h2>
+        </div>
+    </div>
+
+    <div class="card flex-grow-1 rounded-top-5 border-0 card-login">
+        <div class="card-body">
+            <h3 class="fw-bold mb-5 ml" style="color: #600806;">Login</h3>
+
+            <form [formGroup]="loginForm">
+                <div class="form-floating mb-3">
+                    <input type="email" class="form-control" id="userInput" formControlName="user" placeholder="">
+                    <label for="userInput">Usuário</label>
+
+                    <div *ngIf="loginForm.get('user')?.touched && loginForm.get('user')?.invalid"
+                        class="text-danger mt-1 small">
+                        <div *ngIf="loginForm.get('user')?.hasError('required')">Usuário é obrigatório.</div>
+                    </div>
+
+                </div>
+                <div class="form-floating">
+                    <input [type]="hidePassword ? 'text' : 'password'" class="form-control" id="passwordInput"
+                        formControlName="password" placeholder="">
+                    <label for="passwordInput">Senha</label>
+
+
+                    <div *ngIf="loginForm.get('password')?.touched && loginForm.get('password')?.invalid"
+                        class="text-danger mt-1 small">
+                        <div *ngIf="loginForm.get('password')?.hasError('required')">Senha é obrigatória.</div>
+                    </div>
+
+                    <button type="button" *ngIf="btnVisibility"
+                        class="btn btn-link position-absolute top-50 end-0 translate-middle-y me-2 p-0"
+                        (click)="changeVisibility()">
+                        <svg *ngIf="hidePassword" xmlns="http://www.w3.org/2000/svg" width="24" height="24"
+                            fill="currentColor" class="bi bi-eye-slash" viewBox="0 0 16 16" style="color: black;">
+                            <path
+                                d="M13.359 11.238C15.06 9.72 16 8 16 8s-3-5.5-8-5.5a7 7 0 0 0-2.79.588l.77.771A6 6 0 0 1 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13 13 0 0 1 14.828 8q-.086.13-.195.288c-.335.48-.83 1.12-1.465 1.755q-.247.248-.517.486z" />
+                            <path
+                                d="M11.297 9.176a3.5 3.5 0 0 0-4.474-4.474l.823.823a2.5 2.5 0 0 1 2.829 2.829zm-2.943 1.299.822.822a3.5 3.5 0 0 1-4.474-4.474l.823.823a2.5 2.5 0 0 0 2.829 2.829" />
+                            <path
+                                d="M3.35 5.47q-.27.24-.518.487A13 13 0 0 0 1.172 8l.195.288c.335.48.83 1.12 1.465 1.755C4.121 11.332 5.881 12.5 8 12.5c.716 0 1.39-.133 2.02-.36l.77.772A7 7 0 0 1 8 13.5C3 13.5 0 8 0 8s.939-1.721 2.641-3.238l.708.709zm10.296 8.884-12-12 .708-.708 12 12z" />
+                        </svg>
+
+                        <svg *ngIf="!hidePassword" xmlns="http://www.w3.org/2000/svg" width="24" height="24"
+                            fill="currentColor" class="bi bi-eye" viewBox="0 0 16 16" style="color: black;">
+                            <path
+                                d="M16 8s-3-5.5-8-5.5S0 8 0 8s3 5.5 8 5.5S16 8 16 8M1.173 8a13 13 0 0 1 1.66-2.043C4.12 4.668 5.88 3.5 8 3.5s3.879 1.168 5.168 2.457A13 13 0 0 1 14.828 8q-.086.13-.195.288c-.335.48-.83 1.12-1.465 1.755C11.879 11.332 10.119 12.5 8 12.5s-3.879-1.168-5.168-2.457A13 13 0 0 1 1.172 8z" />
+                            <path
+                                d="M8 5.5a2.5 2.5 0 1 0 0 5 2.5 2.5 0 0 0 0-5M4.5 8a3.5 3.5 0 1 1 7 0 3.5 3.5 0 0 1-7 0" />
+                        </svg>
+                    </button>
+                </div>
+
+                <a href="#" class="text-muted small d-block text-start mb-3">Esqueceu a senha?</a>
+                
+                <div *ngIf="loginForm.invalid" class="text-danger mt-1 small">
+                    <div *ngIf="loginForm.getError('infoInvalid')">Usuário ou senha inválidos</div>
+                </div>
+
+                <button class="btn btn-danger w-100 fw-bold py-2" (click)="onSubmit()">Entrar</button>
+                <div class="b-line"></div>
+            </form>
+        </div>
+    </div>
+</div>
